Image shearing
Image shearing is a geometric transformation that skews an image along one or both axes i.e x or y axis.
To shear an image using OpenCV, we need to create a transformation matrix. This matrix is a 2×3 matrix that specifies the amount of shearing in each direction.
The cv2.warpAffine() function is used to apply a transformation matrix to an image. It takes the following arguments:
The image to be transformed.
The transformation matrix.