Edge detection of Image
The process of image edge detection involves detecting sharp edges in the image. This edge detection is essential in the context of image recognition or object localization/detection. There are several algorithms for detecting edges due to its wide applicability.
In image processing and computer vision applications, Canny Edge Detection is a well-liked edge detection approach. In order to detect edges, the Canny edge detector first smoothes the image to reduce noise, then computes its gradient, and then applies a threshold to the gradient. The multi-stage Canny edge detection method includes the following steps:
Gaussian smoothing: The image is smoothed using a Gaussian filter to remove noise.
Gradient calculation: The gradient of the image is calculated using the Sobel operator.
Non-maximum suppression: Non-maximum suppression is applied to the gradient image to remove spurious edges.
Hysteresis thresholding: Hysteresis thresholding is applied to the gradient image to identify strong and weak edges.